feat: add a template for turbo module (experimental)

+++ packages/create-react-native-library/templates/cpp-library/cpp/{%- project.identifier %}.cpp
@@ -1,7 +1,7 @@
-#include "<%- project.identifier %>.h"
+#include "<%- project.identifier -%>.h"
 
 namespace example {
-	int multiply(float a, float b) {
+	int multiply(double a, double b) {
 		return a * b;
 	}
 }
